More adjustments

+++ api/src/main/java/com/github/skjolber/packing/api/ep/Point3D.java
@@ -611,7 +611,7 @@ public abstract class Point3D<P extends Placement3D> extends Point2D<P> {
 	public abstract Point3D<P> rotate();
 
 	public long getVolumeAtZ(int zz) {
-		return (long)dx *  (long)dy * (long)(maxZ - zz + 1);
+		return (long)dx * (long)dy * (maxZ - zz + 1);
 	}
 
 }
